Make a new type internal only
